requirements.txt
��asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
config==0.5.1
cryptography==46.0.3
Django==6.0.1
djangorestframework==3.16.1
dnspython==2.8.0
dotenv==0.9.9
et_xmlfile==2.0.0
h11==0.16.0
idna==3.11
jwt==1.4.0
numpy==2.4.1
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.3
pycparser==2.23
PyJWT==2.10.1
pymongo==4.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
resend==2.19.0
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlparse==0.5.5
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
